pyproject.toml
[tool.poetry]
name = "pars-yandex"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
selenium = "^4.24.0"
webdriver-manager = "^4.0.2"
aiogram = "^3.13.0"
loguru = "^0.7.2"
selenium-stealth = "^1.0.6"
cloudscraper = "^1.2.71"
fake-useragent = "^1.5.1"
yadisk = "^3.1.0"
fast-bitrix24 = "^1.8.0"
gspread = "^6.1.2"
oauth2client = "^4.1.3"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
numpy = "^2.1.3"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
